Address reviewer comments. Don't create a CommonUtils file, but move the code to CommonArgs instead. That code has been put in the clang::driver namespace.
